Make the BukkitChat (aka defaultchat) more advanced.
For one thing, in the config.yml it can be specified if the players listen or ignore by default. If the whitelist option is set to true, a whitelist is maintained, meaning that by default, players ignore auction broadcasts until they decide to use "/auc listen".
The second change is that the list of players who are listening or ignoring, respectively, is saved to disk to store it across restarts. The file is called "defaultchat.yml".